kauai/tools/chomp.c: fix options parser for filenames beginning with /
The existing options parser doesn't work correctly if filenames begin with a
forward slash since it is assumed that anything following the forward slash
will always be a command line option.

Adjust the parser so that only arguments that begin with a forward slash and
have a length of 2 characters are considered to be command line options.