Reduce duplicate mouse tracking messages
Previously, when a GOB was tracking the mouse, cidTrackMouse messages
would be dispatched as fast as they could be generated. This caused
high CPU usage when dragging actors around the stage.

This commit changes the behaviour of CEX::_TGetNextCmd to send
cidTrackMouse commands only if the mouse state has changed.